apps/ai-engine/requirements.txt
# ===============================
# FASTAPI CORE
# ===============================
fastapi==0.112.1
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# ===============================
# DATABASE & MEMORY
# ===============================
pymongo==4.8.0
motor==3.5.1

# IMPORTANT:
# arq DOES NOT support redis 5.x
redis==4.6.0

# ===============================
# VECTOR DB + EMBEDDINGS (RAG)
# ===============================
qdrant-client==1.11.0
fastembed==0.3.1

# ===============================
# AI & LLM CLIENTS
# ===============================
groq==0.10.0
openai==1.40.0
google-generativeai==0.7.2

# Instructor REQUIRES tenacity >=8.4.1,<9
instructor==1.3.7
tenacity==8.5.0

# Voice (optional, keep only if used)
hume==0.6.0

# ===============================
# BACKGROUND WORKERS
# ===============================
arq==0.26.0

# ===============================
# UTILS & SECURITY
# ===============================
httpx==0.27.0
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pypdf==4.2.0
